Msform.McpBootstrap (msapp_mcp v0.1.0)

Form data used during the MCP Bootstrapping process.

Note that currently the data assumes that only an email/password Authenticator will be provided for the administrative Access Account.

Types

t()

@type t() :: %Msform.McpBootstrap{
  admin_credential: String.t() | nil,
  admin_credential_verify: String.t() | nil,
  admin_display_name: String.t() | nil,
  admin_identifier: String.t() | nil,
  id: term(),
  owner_display_name: String.t() | nil,
  owner_name: String.t() | nil
}

Functions

enter_disallowed_state(socket)

See Msform.McpBootstrap.Actions.enter_disallowed_state/1.

enter_finished_state(socket)

See Msform.McpBootstrap.Actions.enter_finished_state/1.

enter_records_state(socket)

See Msform.McpBootstrap.Actions.enter_records_state/1.

enter_welcome_state(socket)

See Msform.McpBootstrap.Actions.enter_welcome_state/1.

finish_processing_override(socket_or_assigns, override)

See MscmpSystForms.finish_processing_override/2.

get_component_info(form_id)

process_disallowed_finished(socket)

See Msform.McpBootstrap.Actions.process_disallowed_finished/1.

process_disallowed_load(socket)

See Msform.McpBootstrap.Actions.process_disallowed_load/1.

process_records_save(socket)

See Msform.McpBootstrap.Actions.process_records_save/1.

process_records_save_finished(socket)

See Msform.McpBootstrap.Actions.process_records_save_finished/1.

start_processing_override(socket_or_assigns, override)

See MscmpSystForms.start_processing_override/2.

update_button_state(socket_or_assigns, form_id, button_state)

See MscmpSystForms.update_button_state/3.

update_display_data(socket_or_assigns, display_data, opts \\ [])

See MscmpSystForms.update_display_data/3.

validate_form_data(socket, form_data)

See Msform.McpBootstrap.Actions.validate_form_data/2.

validate_save(original_data, current_data \\ %{})

@spec validate_save(t(), Msform.McpBootstrap.Types.parameters()) :: Ecto.Changeset.t()

Validates a map of form data for validation of form entry prior to final submission.